Added PlayerStats script
This provides a more concise way to deal with the player health and fire
health. I did have to make many updates to other scripts to get the fire
health working.

+++ Assets/Scripts/RechargeFire.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class RechargeFire : MonoBehaviour {
 
-    public float healAmount = 1;
+    public int healAmount = 1;
 
     void OnTriggerStay2D(Collider2D col)
     {
